+++ examples/publisher/index.ts
@@ -7,6 +7,7 @@ import prebid from 'prebid.js';
 import 'prebid.js/modules/consentManagement';
 import 'prebid.js/modules/currency';
 import 'prebid.js/modules/appnexusBidAdapter';
+import 'prebid.js/modules/justpremiumBidAdapter';
 import 'prebid.js/modules/pubmaticBidAdapter';
 import 'prebid.js/modules/smartadserverBidAdapter';
 import 'prebid.js/modules/teadsBidAdapter';
@@ -18,10 +19,16 @@ import { initAdTag } from '@highfivve/ad-tag';
 import { adConfiguration } from './source/ts/configuration';
 import Cmp from '@highfivve/module-cmp-faktor';
 import Confiant from '@highfivve/module-confiant';
+// import Skin from '@highfivve/module-generic-skin';
 
 prebid.processQueue();
 
 const moli = initAdTag(window);
+
+// skin
+// moli.registerModule(new Skin({
+//   wallpaperAdSlotDomId: 'gf_header'
+// }));
 
 // ad fraud protection
 moli.registerModule(new Confiant({